Move FormTokenField to editor components
[PREMIUM-213]
