+++ mailpoet/assets/js/src/automation/editor/components/form-token-field/index.tsx
