+++ mailpoet/assets/js/src/automation/editor/components/index.ts
@@ -1,1 +1,2 @@
 export * from './panel';
+export * from './form-token-field';
