--- mailpoet/assets/js/src/automation/integrations/components/index.ts
@@ -1,1 +1,0 @@
-export * from '../mailpoet/components/form-token-field';
