+++ mailpoet/assets/js/src/automation/integrations/mailpoet/steps/wp-user-registered/edit/role.ts
@@ -1,4 +1,4 @@
-import { FormTokenItem } from '../../../components/form-token-field';
+import { FormTokenItem } from '../../../../../editor/components';
 
 declare global {
   interface Window {
