+++ mailpoet/assets/js/src/webpack_admin_expose.js
@@ -30,4 +30,3 @@ export * as Hooks from 'hooks.js';
 export * as Listing from 'listing';
 export * as DynamicSegmentsTypes from 'segments/dynamic/types';
 export * as AutomationEditorComponents from 'automation/editor/components';
-export * as AutomationIntegrationComponents from 'automation/integrations/components';
